+++ lib/school/class2nd/service/activity_details.dart
@@ -3,7 +3,7 @@ import 'dart:collection';
 import 'package:beautiful_soup_dart/beautiful_soup.dart';
 import 'package:intl/intl.dart';
 import 'package:mimir/network/session.dart';
-import 'package:mimir/session/sc.dart';
+import 'package:mimir/session/class2nd.dart';
 
 import '../entity/details.dart';
