+++ lib/school/class2nd/service/score.dart
@@ -2,7 +2,7 @@ import 'package:beautiful_soup_dart/beautiful_soup.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:intl/intl.dart';
 import 'package:mimir/network/session.dart';
-import 'package:mimir/session/sc.dart';
+import 'package:mimir/session/class2nd.dart';
 
 import '../entity/list.dart';
 import '../entity/score.dart';
