+++ lib/session/common.dart
@@ -3,17 +3,6 @@ import 'package:mimir/network/session.dart';
 
 import '../network/download.dart';
 
-extension DioResTypeConverter on ResponseType {
-  ResponseType toDioResponseType() {
-    return (const {
-      ResponseType.json: ResponseType.json,
-      ResponseType.stream: ResponseType.stream,
-      ResponseType.plain: ResponseType.plain,
-      ResponseType.bytes: ResponseType.bytes,
-    })[this]!;
-  }
-}
-
 extension DioOptionsConverter on SessionOptions {
   Options toDioOptions() {
     return Options(
@@ -22,7 +11,7 @@ extension DioOptionsConverter on SessionOptions {
       receiveTimeout: receiveTimeout,
       extra: extra,
       headers: headers,
-      responseType: responseType?.toDioResponseType(),
+      responseType: responseType,
       contentType: contentType,
     );
   }
